Duplicate bugs removed

+++ src/main/java/com/csi/dao/impl/EmployeeDaoImpl.java
@@ -39,7 +39,9 @@ public class EmployeeDaoImpl implements EmployeeDao {
 
     @Override
     public void deleteEmployeeById(int empId) {
-        employeeRepository.deleteById(empId);
+        if (employeeRepository.existsById(empId)) {
+            employeeRepository.deleteById(empId);
+        }
     }
 
     @Override
